+++ src/main/java/com/example/blog5/service/UserService.java
@@ -1,8 +1,10 @@
 package com.example.blog5.service;
 
+import com.example.blog5.model.Blog;
 import com.example.blog5.model.User;
 import com.example.blog5.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -18,9 +20,13 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+
     public User findByUsername(String username) {
         return userRepository.findByUsername(username);
     }
+
     public User getUserById(Long id) {
         return userRepository.findById(id).orElse(null);
     }
@@ -53,7 +59,31 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(User user) {
-        userRepository.delete(user);
+    public boolean deleteUserWithPasswordCheck(Long userId, String password) {
+        User user = userRepository.findById(userId).orElse(null);
+        if (user != null && passwordEncoder.matches(password, user.getPassword())) {
+            userRepository.delete(user);
+            return true;
+        }
+        return false;
+    }
+
+    public boolean updateUser(Long userId, String name, String email, String profileImage, String blogTitle, Map<String, Boolean> emailNotification) {
+        User user = userRepository.findById(userId).orElse(null);
+        if (user != null) {
+            user.setUsername(name);
+            user.setEmail(email);
+            user.setProfileImage(profileImage);
+
+            Blog blog = user.getBlog();
+            if (blog != null) {
+                blog.setTitle(blogTitle);
+            }
+
+            user.setEmailNotifications(emailNotification);
+            userRepository.save(user);
+            return true;
+        }
+        return false;
     }
 }
